Improved install script

+++ install.sh
@@ -1,4 +1,25 @@
 #!/bin/bash
+set -e
 
-sudo cp src/pug.sh /usr/local/bin/pug
-cp src/installers/* ~/.pug/installers
+install_from_repo() {
+  echo "Installing pug"
+  if [ "$1" != -l ]; then
+    echo 'Password required to copy pug to /usr/local/bin/pug'
+    sudo cp src/pug.sh /usr/local/bin/pug
+  fi
+
+  echo 'Copying installers to ~/.pug/installers'
+  mkdir -p ~/.pug/installers
+  cp src/installers/* ~/.pug/installers
+}
+
+if [ -e ./src/pug.sh ]; then
+  echo 'pug.sh exists, installing locally'
+  install_from_repo "$@"
+else
+  echo 'pug.sh does not exist, cloning to /tmp'
+  cd /tmp
+  git clone 'https://github.com/javanut13/pug.git'
+  cd pug
+  install_from_repo
+fi
